# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing shared expenses, groups, and user authentication",
        "title": "Qashare API",
        "contact": {
            "name": "Pranaov S",
            "email": "qashare.contact@pranaovs.me"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/groups/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new group with the logged in user as the creator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Create a new group",
                "parameters": [
                    {
                        "description": "Group details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "description": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Group successfully created",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body format or missing required fields | BAD_NAME: Name contains invalid characters or is too short/long",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
        "/v1/groups/admin": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all groups that the authenticated user created (is admin of)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List groups user administers",
                "deprecated": true,
                "responses": {
                    "200": {
                        "description": "Returns list of groups the user is admin of",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Group"
                            }
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
        "/v1/groups/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all groups the logged in user is a member of",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List user's groups",
                "deprecated": true,
                "responses": {
                    "200": {
                        "description": "Returns list of groups the user is a member of",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Group"
                            }
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
        "/v1/groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get group details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns group details including members and expenses",
                        "schema": {
                            "$ref": "#/definitions/models.GroupDetails"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "USERS_NOT_RELATED: The authenticated user is not a member of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "GROUP_NOT_FOUND: The specified group does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update group name and description (requires group admin permission). Immutable fields will be ignored if included in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Update a group (full replacement)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated group details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated group",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "NO_PERMISSIONS: User is not the group admin | USERS_NOT_RELATED: The authenticated user is not a member of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "GROUP_NOT_FOUND: The specified group does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update specific fields of a group. Only provided fields are updated, others remain unchanged.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Partially update a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial group details (name and/or description, all optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated group with all fields",
                        "schema": {
                            "$ref": "#/definitions/models.GroupDetails"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "NO_PERMISSIONS: User is not the group admin | USERS_NOT_RELATED: The authenticated user is not a member of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "GROUP_NOT_FOUND: The specified group does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
        "/v1/groups/{id}/expenses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all expenses of a group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List group expenses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of all expenses in the group",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Expense"
                            }
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "USERS_NOT_RELATED: The authenticated user is not a member of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "GROUP_NOT_FOUND: The specified group does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierrors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g., \"BAD_NAME\", \"INVALID_EMAIL\"",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                }
            }
        },
...
        "models.Group": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.GroupDetails": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GroupUser"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}